Changed Integer to Long in code generation. Fixed problem with single type names and single quoted values in shorthand definitions. Bootstrapped treegen to use latest version of abstract syntax tree classes.
git-svn-id: https://overture.svn.sourceforge.net/svnroot/overture/trunk@2190 2b2d4d09-c93c-414a-ad04-6e973f5c1cff

+++ tools/treegen/src/main/java/org/overture/tools/treegen/ast/imp/TreeGenAstNode.java
@@ -21,31 +21,31 @@ public class TreeGenAstNode implements ITreeGenAstNode
 	}
 
 	// private member variable (column)
-	private Integer m_column;
+	private Long m_column;
 
 	// public set operation for private member variable (column)
-	public void setColumn(Integer piv)
+	public void setColumn(Long piv)
 	{
 		m_column = piv;
 	}
 
 	// public get operation for private member variable (column)
-	public Integer getColumn()
+	public Long getColumn()
 	{
 		return m_column;
 	}
 
 	// private member variable (line)
-	private Integer m_line;
+	private Long m_line;
 
 	// public set operation for private member variable (line)
-	public void setLine(Integer piv)
+	public void setLine(Long piv)
 	{
 		m_line = piv;
 	}
 
 	// public get operation for private member variable (line)
-	public Integer getLine()
+	public Long getLine()
 	{
 		return m_line;
 	}
